fix(serverless): Use ingest_stream for Axiom logs
This fixes a bug where only the very first event would've been sent to
Axiom.
The receiver is now transformed into an async stream and passed to
`ingest_stream`, which will automatically send events every second or in
batches of 1000 (whichever happens first).

It also renames the `timestamp` field to `_time` so Axiom recognizes it.

And it upgrades axiom-rs to 0.6 because that release added the flush
every second functionality.